requirements.txt
psycopg2-binary
dj-database-url
djangoajax
whitenoise
djangorestframework
markdown
Pillow